requirements.txt
psycopg2-binary
requests
tenacity
python-dotenv
pyyaml